requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
pillow==10.4.0
openai==1.52.2
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
ruff==0.6.9
